requirements.txt
argparse
coverage
cryptography
datasets
google-api-python-client
grpcio
grpcio-tools
loguru
numpy
pickle-mixin
pycryptodome
pycrypto
pytest
tensorboard
torch
torchvision
transformers